requirements.txt
certifi==2019.6.16
chardet==3.0.4
future==0.17.1
idna==2.8
oauthlib==3.0.2
pkg-resources==0.0.0
python-twitter==3.5
requests==2.22.0
requests-oauthlib==1.2.0
urllib3==1.25.3
